Add files via upload
JS files, but logic are yet to be developed, this is eating me and my brain cells, I am a bit poor at developing logic right now, even a small terminology hits me so well(sometimes i think  this programming is not for me), but i will work on it and i will make sure the logic of the calculator is fully developed, i will keep on updating the source codes of the web-app. Thanks! Happy coding.